Refactor application structure to service-based model
Restructured the code to shift from a traditional application-based structure to a services-based model in order to enhance modularity and scalability. Deleted internal/application/application.go as it was redundant and replaced by internal/service/service.go. Similarly, service_test.go was also removed. All occurrences of "application" were replaced with the term "service" in root.go, main.go, version.go, client.go and README.md files. No change in business logic and the code should behave the same as before. This refactor is to prepare for further code enhancements and easier manageability.