Refactor command structures for better encapsulation
Refactored the `BuildCommand` and `BuildOptions` into a simpler and more encapsulated structure known as `State` in 'internal/command/command.go'. This will enhance readability, maintainability and scalability of the source code due to the clearer separation of concerns. It also effectively simplifies the process of adding new command flags.

The new 'State' structure instead of 'BuildOptions' has been integrated into the system in 'internal/service/service.go' replacing the map with a slice of 'State'. This structural adjustment further supports better code organization and modification in the future.